+++ src/variables/SprigVariable.php
@@ -7,7 +7,6 @@ namespace putyourlightson\sprig\variables;
 
 use Craft;
 use craft\db\Paginator;
-use craft\helpers\Html;
 use craft\helpers\Template;
 use craft\web\twig\variables\Paginate;
 use putyourlightson\sprig\base\Component;
@@ -18,22 +17,25 @@ use yii\db\Query;
 class SprigVariable
 {
     /**
-     * @var string The htmx version to load (must exist in `src/resources/lib/htmx`).
-     * Downloaded from https://unpkg.com/htmx.org
-     */
-    public string $htmxVersion = '1.9.2';
-
-    /**
      * Returns the script tag with the given attributes.
+     *
+     * @deprecated in 2.6.0
      */
     public function getScript(array $attributes = []): Markup
     {
-        $path = '@putyourlightson/sprig/resources/lib/htmx/' . $this->htmxVersion . '/';
-        $path .= Craft::$app->getConfig()->env == 'dev' ? 'htmx.js' : 'htmx.min.js';
-        $url = Craft::$app->getAssetManager()->getPublishedUrl($path, true);
-        $script = Html::jsFile($url, $attributes);
+        Craft::$app->getDeprecator()->log(__METHOD__, '`sprig.script` has been deprecated. It is no longer required and can be safely removed.');
 
-        return Template::raw($script);
+        return Template::raw('');
+    }
+
+    /**
+     * Returns the URL to the script file.
+     *
+     * @added in 2.6.0
+     */
+    public function getScriptUrl(): string
+    {
+        return Sprig::$core->components->getScriptUrl();
     }
 
     /**
@@ -171,6 +173,7 @@ class SprigVariable
 
     /**
      * Returns a [[RefreshOnLoad]] component.
+     *
      * @see https://github.com/putyourlightson/craft-sprig/issues/279
      */
     public function triggerRefreshOnLoad(string $selector = null): Markup