+++ tests/unit/services/ComponentsTest.php
@@ -10,6 +10,7 @@ use Craft;
 use craft\elements\Entry;
 use putyourlightson\sprig\services\ComponentsService;
 use putyourlightson\sprig\Sprig;
+use Twig\Markup;
 use UnitTester;
 use yii\base\Application;
 use yii\base\ExitException;
@@ -22,7 +23,6 @@ use yii\web\Request;
  * @package   Sprig
  * @since     1.0.0
  */
-
 class ComponentsTest extends Unit
 {
     protected UnitTester $tester;
@@ -38,6 +38,20 @@ class ComponentsTest extends Unit
         Sprig::bootstrap();
 
         Craft::$app->getView()->setTemplatesPath(Craft::getAlias('@templates'));
+    }
+
+    public function testHtmxScriptExistsForDev()
+    {
+        Craft::$app->getConfig()->env = 'dev';
+
+        $this->_testScriptExistsLocally();
+    }
+
+    public function testHtmxScriptExistsForProduction()
+    {
+        Craft::$app->getConfig()->env = 'production';
+
+        $this->_testScriptExistsLocally();
     }
 
     public function testCreate()
@@ -118,9 +132,14 @@ class ComponentsTest extends Unit
         $this->assertNull($object);
     }
 
-    public function testCreateInvalidVariableEntry()
+    public function testCreateVariableEntry()
     {
-        $this->_testCreateInvalidVariable(['number' => '', 'entry' => new Entry()]);
+        $this->_testCreateVariable(['number' => '', 'entry' => new Entry()]);
+    }
+
+    public function testCreateValidVariableArray()
+    {
+        $this->_testCreateVariable(['number' => '', 'array' => [new Entry()]]);
     }
 
     public function testCreateInvalidVariableModel()
@@ -135,7 +154,7 @@ class ComponentsTest extends Unit
 
     public function testCreateInvalidVariableArray()
     {
-        $this->_testCreateInvalidVariable(['number' => '', 'array' => [new Entry()]]);
+        $this->_testCreateInvalidVariable(['number' => '', 'array' => [new Model()]]);
     }
 
     public function testGetParsedTagAttributes()
@@ -276,6 +295,24 @@ class ComponentsTest extends Unit
         $this->assertStringContainsString($placeholder, $result);
     }
 
+    private function _testScriptExistsLocally(): void
+    {
+        $url = Sprig::$core->components->getScriptUrl();
+        preg_match('/cpresources(.*?)\?v=/', $url, $matches);
+        $path = Craft::getAlias(Craft::$app->getConfig()->getGeneral()->resourceBasePath) . $matches[1];
+
+        $this->assertFileExists($path);
+    }
+
+    private function _testCreateVariable(array $variables): void
+    {
+        $this->tester->mockCraftMethods('view', ['doesTemplateExist' => true]);
+        Craft::$app->getView()->setTemplatesPath(Craft::getAlias('@templates'));
+        $markup = Sprig::$core->components->create('_component', $variables);
+
+        $this->assertInstanceOf(Markup::class, $markup);
+    }
+
     private function _testCreateInvalidVariable(array $variables): void
     {
         $this->tester->mockCraftMethods('view', ['doesTemplateExist' => true]);
@@ -283,6 +320,7 @@ class ComponentsTest extends Unit
 
         /**
          * Yii exits with an exception when `YII_ENV_TEST` is set.
+         *
          * @see Application::end()
          */
         $this->expectException(ExitException::class);